src/requirements.txt
langchain
langchain-openai
langchain-tavily
langchain-experimental
langgraph
pydantic
pyyaml
jupyter
matplotlib
trulens-core
trulens-dashboard
trulens-apps-langgraph
trulens-providers-openai
numpy
openai
"langserve[all]"